update to snapshot spandsp-20080911.tar.gz
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@9767 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/spandsp/src/dds_int.c
@@ -22,7 +22,7 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- * $Id: dds_int.c,v 1.10 2008/09/01 16:07:33 steveu Exp $
+ * $Id: dds_int.c,v 1.11 2008/09/11 15:13:42 steveu Exp $
  */
 
 /*! \file */
@@ -252,11 +252,11 @@ int16_t dds(uint32_t *phase_acc, int32_t phase_rate)
 }
 /*- End of function --------------------------------------------------------*/
 
-int16_t dds_mod(uint32_t *phase_acc, int32_t phase_rate, int scale, int32_t phase)
+int16_t dds_mod(uint32_t *phase_acc, int32_t phase_rate, int16_t scale, int32_t phase)
 {
     int16_t amp;
 
-    amp = (int16_t) ((dds_lookup(*phase_acc + phase)*scale) >> 15);
+    amp = (int16_t) (((int32_t) dds_lookup(*phase_acc + phase)*(int32_t) scale) >> 15);
     *phase_acc += phase_rate;
     return amp;
 }
@@ -278,12 +278,12 @@ complexi_t dds_complexi(uint32_t *phase_acc, int32_t phase_rate)
 }
 /*- End of function --------------------------------------------------------*/
 
-complexi_t dds_complexi_mod(uint32_t *phase_acc, int32_t phase_rate, int scale, int32_t phase)
+complexi_t dds_complexi_mod(uint32_t *phase_acc, int32_t phase_rate, int16_t scale, int32_t phase)
 {
     complexi_t amp;
 
-    amp = complex_seti((dds_lookup(*phase_acc + phase + (1 << 30))*scale) >> 15,
-                       (dds_lookup(*phase_acc + phase)*scale) >> 15);
+    amp = complex_seti(((int32_t) dds_lookup(*phase_acc + phase + (1 << 30))*(int32_t) scale) >> 15,
+                       ((int32_t) dds_lookup(*phase_acc + phase)*(int32_t) scale) >> 15);
     *phase_acc += phase_rate;
     return amp;
 }
@@ -305,12 +305,12 @@ complexi16_t dds_complexi16(uint32_t *phase_acc, int32_t phase_rate)
 }
 /*- End of function --------------------------------------------------------*/
 
-complexi16_t dds_complexi16_mod(uint32_t *phase_acc, int32_t phase_rate, int scale, int32_t phase)
+complexi16_t dds_complexi16_mod(uint32_t *phase_acc, int32_t phase_rate, int16_t scale, int32_t phase)
 {
     complexi16_t amp;
 
-    amp = complex_seti16((dds_lookup(*phase_acc + phase + (1 << 30))*scale) >> 15,
-                         (dds_lookup(*phase_acc + phase)*scale) >> 15);
+    amp = complex_seti16(((int32_t) dds_lookup(*phase_acc + phase + (1 << 30))*(int32_t) scale) >> 15,
+                         ((int32_t) dds_lookup(*phase_acc + phase)*(int32_t) scale) >> 15);
     *phase_acc += phase_rate;
     return amp;
 }
@@ -332,12 +332,12 @@ complexi32_t dds_complexi32(uint32_t *phase_acc, int32_t phase_rate)
 }
 /*- End of function --------------------------------------------------------*/
 
-complexi32_t dds_complexi32_mod(uint32_t *phase_acc, int32_t phase_rate, int scale, int32_t phase)
+complexi32_t dds_complexi32_mod(uint32_t *phase_acc, int32_t phase_rate, int16_t scale, int32_t phase)
 {
     complexi32_t amp;
 
-    amp = complex_seti32((dds_lookup(*phase_acc + phase + (1 << 30))*scale) >> 15,
-                         (dds_lookup(*phase_acc + phase)*scale) >> 15);
+    amp = complex_seti32(((int32_t) dds_lookup(*phase_acc + phase + (1 << 30))*(int32_t) scale) >> 15,
+                         ((int32_t) dds_lookup(*phase_acc + phase)*(int32_t) scale) >> 15);
     *phase_acc += phase_rate;
     return amp;
 }
